added deletion of images on updates

+++ controllers/admin.js
@@ -1,3 +1,4 @@
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
@@ -155,6 +156,8 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        // fire and forget manner
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
 
@@ -187,8 +190,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  //more authorizations
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(
+    prodId.then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
+  )
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
